update(index): update index creation
